refactor: extract useDisplayPreferences hook from AppContext
Move theme + companion display settings (uiTheme, companionVrmPowerMode,
companionAnimateWhenHidden, companionHalfFramerateMode) with their
normalize-on-set wrappers and persist effects into a standalone hook.

AppContext.tsx: 8,048 → 8,013 lines (-35 from this extraction)
useDisplayPreferences.ts: 96 lines (new)

Cumulative AppContext reduction: 8,366 → 8,013 (-353 lines, 4 hooks extracted)
